Addressing review comments; Add member feature temperarily commented out
