+++ frontend/src/modules/Project/CreateNewProjectForm/CreateNewProjectForm.module.scss
@@ -117,6 +117,7 @@ input {
 .actions {
   display: inline-flex;
   align-items: center;
+  margin-top: 20px;
   gap: 16px;
 }
